requirements.txt
django>=3.0.7
# psycopg2==2.8
djangorestframework==3.11
pygments==2.6
django-cors-headers==3.4.0
django-filter==2.3.0
zipfile38==0.0.2
requests>=2.24.0
django-q>=1.3.3
youtube_dl>=2020.12.7